wip get template working
