+++ django_app/report_a_suspected_breach/templates/report_a_suspected_breach/form_steps/complete.html
@@ -21,7 +21,7 @@
         <p class="govuk-body">We have sent you a confirmation email.</p>
     {% endif %}
     <p class="govuk-body">
-        <a href="{% url 'download_report' %}?reference={{ reference }}" class="govuk-link" target="_blank">Download and view report (opens in new tab)</a>
+        <a href="{% url 'report_a_suspected_breach:download_report' %}?reference={{ reference }}" class="govuk-link" target="_blank">Download and view report (opens in new tab)</a>
     </p>
 
 {% comment %}
