+++ django_app/report_a_suspected_breach/templates/report_a_suspected_breach/form_steps/report_pdf.html
@@ -43,47 +43,49 @@
                 Unknown regime<br>
             {% endif %}
             {% if breach.other_sanctions_regime %}
-                Other regime
+                Other regime<br>
             {% endif %}
+            <br>
             <strong>What were the goods or services?</strong> {{ breach.what_were_the_goods }}<br>
         </div>
 
         <h2 class="govuk-heading-m">People and businesses involved</h2>
         <h3 class="govuk-heading-s">Supplier</h3>
-        <div class="govuk-panel__body"><strong>Name and address of supplier:</strong> {{ supplier.name }}
-            <br> <br>
+        <div class="govuk-panel__body"><strong>Name and address of supplier:</strong> <br>{{ supplier.name }}
+            <br>
             {{ supplier_address | linebreaksbr }}</div>
 
-        {% for end_user in end_users %}
-            <h3 class="govuk-heading-s">End-user {{ forloop.counter }}</h3>
-            <div class="govuk-panel__body">
+        <div class="govuk-panel__body">
+            {% for end_user in end_users %}
+                <h3 class="govuk-heading-s">End-user {{ forloop.counter }}</h3>
                 <strong>Name:</strong> {{ end_user.name }}<br>
                 <strong>Business:</strong> {{ end_user.name_of_business }}<br>
-                <strong>Address:</strong> {% with readable_address=end_user.get_readable_address %}{% with readable_address=end_user.get_readable_address %}
+                <strong>Address:</strong> {{ end_user.get_readable_address }}
                 <strong>Additional contact information:</strong> {{ end_user.additional_contact_details }}
+            {% endfor %}
+            <h3 class="govuk-heading-s">Other people or businesses involved in the trade</h3>
+            <strong>Names and addresses:</strong> {{ breach.other_addresses_in_the_supply_chain }}
+        </div>
 
-                <h3 class="govuk-heading-s">Other people or businesses involved in the trade</h3>
-                <strong>Names and addresses:</strong> {{ breach.other_addresses_in_the_supply_chain }}
-            </div>
+        <h2 class="govuk-heading-m">Sanctions breach details</h2>
+        <div class="govuk-panel__body">
+            <strong>Upload supporting documents:</strong>
+            {% if not breach.documents.all %}
+                None uploaded
+            {% else %}
+                {% for document in breach.documents.all %}
+                    <a href="{{ document.url }}" download>{{ document.file_name }}</a><br>
+                {% endfor %}
+            {% endif %}
+            <strong>Summary of the breach:</strong> {{ breach.tell_us_about_the_suspected_breach }}<br>
+        </div>
 
-            <h2 class="govuk-heading-m">Sanctions breach details</h2>
-            <div class="govuk-panel__body">
-                <strong>Upload supporting documents:</strong> {% if not breach.documents.all %}
-                    None uploaded
-                {% else %}
-                    {% for document in breach.documents.all %}
-                        <a href="{{ document.url }}" download>{{ document.file_name }}</a><br>
-                    {% endfor %}
-                {% endif %}
-                <strong>Summary of the breach:</strong> {{ breach.tell_us_about_the_suspected_breach }}<br>
-            </div>
+        <h2 class="govuk-heading-m">What happens next</h2>
+        <div class="govuk-panel__body">
+            We've sent your report to the Office of Trade Sanctions Implementation (OTSI), part of the Department for Business and Trade (DBT).<br>
 
-            <h2 class="govuk-heading-m">What happens next</h2>
-            <div class="govuk-panel__body">
-                We've sent your report to the Office of Trade Sanctions Implementation (OTSI), part of the Department for Business and Trade (DBT).<br>
-
-                OTSI will contact you once they've reviewed your report.
-            </div>
-            </div>
-        {% endblock column_content %}
-        {% block footer %}{% endblock footer %}
+            OTSI will contact you once they've reviewed your report.
+        </div>
+    </div>
+{% endblock column_content %}
+{% block footer %}{% endblock footer %}
