+++ django_app/report_a_suspected_breach/urls.py
@@ -9,6 +9,8 @@ from report_a_suspected_breach.views import (
     views_supply_chain,
 )
 
+from django_app.report_a_suspected_breach.views.views_download import DownloadPDFView
+
 app_name = "report_a_suspected_breach"
 
 generic_urls = [
@@ -123,6 +125,7 @@ views_end_urls = [
     path("check-your-answers", views_end.CheckYourAnswersView.as_view(), name="check_your_answers"),
     path("declaration", views_end.DeclarationView.as_view(), name="declaration"),
     path("submission-complete", views_end.CompleteView.as_view(), name="complete"),
+    path("download_report/", DownloadPDFView.as_view(), name="download_report"),
 ]
 
 urlpatterns = (
